// eventpass-web/src/features/auth/schema.ts
import { z } from "zod";

export const loginSchema = z.object({
    email: z
        .email("Formato de e-mail inválido")
        .min(1, "O e-mail é obrigatório"),
    password: z.string().min(6, "A senha deve ter no mínimo 6 caracteres"),
});

export const registerSchema = z.object({
    name: z.string().min(3, "O nome deve ter no mínimo 3 caracteres"),
    email: z.email("Formato de e-mail inválido"),
    password_encrypted: z
        .string()
        .min(6, "A senha deve ter no mínimo 6 caracteres"),
});

export const registerByAdminSchema = z.object({
    name: z.string().min(3, "O nome deve ter no mínimo 3 caracteres"),
    email: z.email("Formato de e-mail inválido"),
    password_encrypted: z.string().min(6, "A senha deve ter no mínimo 6 caracteres"),
    role: z.string()
});